\name{magicaxis-package}
\alias{magicaxis-package}
\alias{mag}
\docType{package}
\title{
Pretty scientific plotting with minor-tick and log minor-tick support
}
\description{
Functions to make pretty axes (major and minor) on scientific plots. Particularly effort is made on producing nice log plot outputs. The core function produces pretty axis labelling in a number of circumstances that are often used in scientific plotting. There is a higher level interface to a generic plot function that will usually produce nice plots, even without much though on the users part.
}
\details{
\tabular{ll}{
Package: \tab magicaxis\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2013-08-27\cr
License: \tab GPL-2 \cr
Depends: \tab MASS, plotrix \cr
}
}
