\name{magicaxis-package}
\alias{magicaxis-package}
\alias{magicaxis}
\docType{package}
\title{
Pretty Scientific Plotting with Minor-Tick and Log Minor-Tick Support
}
\description{
Functions to make pretty axes (major and minor) on scientific plots. Particularly effort is made on producing nice log plot outputs. The core function produces pretty axis labelling in a number of circumstances that are often used in scientific plotting. There is a higher level interface to a generic plot function that will usually produce nice plots, even without much though on the users part.
}
\details{
\tabular{ll}{
Package: \tab magicaxis\cr
Type: \tab Package\cr
Version: \tab 2.0.1\cr
Date: \tab 2017-05-11\cr
License: \tab GPL-3 \cr
Depends: \tab MASS, plotrix, sm, mapproj, celestial \cr
}
}
